+++ apps/view/src/components/chart-components.tsx
@@ -1,9 +1,11 @@
+import { clamp } from '@dash/common';
 import { motion } from 'framer-motion';
 import { FC, useMemo, useRef, useState } from 'react';
 import {
   Area,
   AreaChart,
   BarChart,
+  LabelProps,
   Pie,
   PieChart,
   Sector,
@@ -231,6 +233,31 @@ const ToolTipContainer = styled.div`
   gap: 5px;
 `;
 
+export const VertBarStartLabel: FC<
+  LabelProps & {
+    labelRenderer: (value: number) => string;
+  }
+> = props => {
+  const theme = useTheme();
+
+  return (
+    <text
+      x={props.x}
+      y={(props.y as number) + (props.height as number) / 2}
+      width={props.width}
+      height={props.height}
+      offset={props.offset}
+      className='recharts-text recharts-label'
+      text-anchor='start'
+      fill={theme.colors.text}
+    >
+      <tspan x={(props.x as number) + (props.offset as number)} dy='0.355em'>
+        {props.labelRenderer(props.value as number)}
+      </tspan>
+    </text>
+  );
+};
+
 type DefaultVertBarChartProps = {
   height: number;
   width: number;
@@ -246,7 +273,7 @@ export const DefaultVertBarChart: FC<DefaultVertBarChartProps> = ({
   children,
   tooltipRenderer,
 }) => {
-  const barSize = Math.min(Math.max(height / data.length - 20, 20), 50);
+  const barSize = clamp(height / data.length - 20, 10, 60);
   const gap = (data.length - 1) * 10;
   const allBars = barSize * data.length;
   const margin = (height - gap - allBars) / 2;
@@ -259,7 +286,7 @@ export const DefaultVertBarChart: FC<DefaultVertBarChartProps> = ({
       layout={'vertical'}
       margin={{
         top: margin,
-        bottom: margin,
+        bottom: data.length === 1 ? margin / 2 : margin,
         right: 20,
         left: 20,
       }}